Add files via upload
Sponsorship proposal or business development folders are a creative tool that I utilize to captivate the minds of potential clients. I have had extensive experience in creating these folders as part of my involvement with the ECC Bureau des Sports, where we regularly organize events and seek sponsorship. These folders are carefully crafted to include various elements such as event regulations, invitations, badges, event posters, and event structure. The goal is to create a visually appealing and comprehensive package that effectively communicates the details and benefits of the event, enticing sponsors to participate and support our initiatives. Through my experience, I have developed skills in designing these folders to be engaging and persuasive, showcasing the value and opportunities of the events we organize to potential sponsors and participants.